pyproject.toml
[tool.poetry]
name = "lct-24"
version = "0.1.0"
description = ""
authors = ["kirill <shkolenko02-032-2004@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
psycopg2-binary = "^2.9.9"
openpyxl = "^3.1.3"
numpy = "^1.26.4"
sqlalchemy = "^2.0.30"
loguru = "^0.7.2"
geoalchemy2 = "^0.15.1"
dash = "^2.17.0"
dash-auth = "^2.3.0"
dash-leaflet = "^1.0.15"
apscheduler = "^3.10.4"
flask-caching = "^2.3.0"
dash-extensions = "^1.0.16"
folium = "^0.16.0"
geopandas = "^0.14.4"
fastapi = {extras = ["all"], version = "^0.111.0"}
plotly = "^5.22.0"
geojson = "^3.1.0"
pyshp = "^2.3.1"
alembic = "^1.13.1"
ruff = "^0.4.9"
pre-commit = "^3.7.1"
fiona = "^1.9.6"
reportlab = "^4.2.0"
pyproj = "^3.6.1"
ultralytics = "^8.2.32"
pillow = "^10.3.0"
opencv-python = "^4.10.0.82"
rasterio = "^1.3.10"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
